code fixes

+++ src/main/java/com/utku/exchange/data/dto/response/ExchangeResultDto.java
@@ -3,6 +3,9 @@ package com.utku.exchange.data.dto.response;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
+
+import java.math.BigDecimal;
+
 /**
  * @author Utku APAYDIN
  * @created 17/05/2022 - 20:50
@@ -11,6 +14,6 @@ import lombok.Data;
 @Builder
 @AllArgsConstructor
 public class ExchangeResultDto {
-    private Double calculatedAmount;
+    private BigDecimal calculatedAmount;
     private String transactionId;
 }
