+++ src/test/java/com/utku/exchange/data/MockTestData.java
@@ -2,6 +2,7 @@ package com.utku.exchange.data;
 
 import com.utku.exchange.data.entity.ExchangeHistory;
 
+import java.math.BigDecimal;
 import java.text.ParseException;
 import java.text.SimpleDateFormat;
 import java.util.*;
@@ -32,11 +33,11 @@ public class MockTestData {
     }
     public static List<ExchangeHistory> getExchangeHistoryData() throws ParseException {
         List<ExchangeHistory> data = new ArrayList<>();
-        data.add(new ExchangeHistory("1","transaction-1","JEP","AMD", 1.5,100.1, FORMATTER.parse("2022-01-10")));
-        data.add(new ExchangeHistory("2","transaction-2","BRL","BDT",100.2,2.5,  FORMATTER.parse("2022-02-15")));
-        data.add(new ExchangeHistory("3","transaction-3","GTQ","BMD",100.3,1.1,  FORMATTER.parse("2022-03-20")));
-        data.add(new ExchangeHistory("4","transaction-4","GTQ","CVE",100.4,5.0,  FORMATTER.parse("2022-04-10")));
-        data.add(new ExchangeHistory("5","transaction-5","JEP","BMD",100.5,1.2,  FORMATTER.parse("2022-05-11")));
+        data.add(new ExchangeHistory("1","transaction-1","JEP","AMD", BigDecimal.valueOf(1.5), BigDecimal.valueOf(100.1), FORMATTER.parse("2022-01-10")));
+        data.add(new ExchangeHistory("2","transaction-2","BRL","BDT", BigDecimal.valueOf(100.2), BigDecimal.valueOf(2.5),  FORMATTER.parse("2022-02-15")));
+        data.add(new ExchangeHistory("3","transaction-3","GTQ","BMD", BigDecimal.valueOf(100.3), BigDecimal.valueOf(1.1),  FORMATTER.parse("2022-03-20")));
+        data.add(new ExchangeHistory("4","transaction-4","GTQ","CVE", BigDecimal.valueOf(100.4), BigDecimal.valueOf(5.0),  FORMATTER.parse("2022-04-10")));
+        data.add(new ExchangeHistory("5","transaction-5","JEP","BMD", BigDecimal.valueOf(100.5), BigDecimal.valueOf(1.2),  FORMATTER.parse("2022-05-11")));
         return data;
     }
 
@@ -51,65 +52,65 @@ public class MockTestData {
         return data;
     }
 
-    public static Map<String,Double> getRateForSymbol(String symbol){
+    public static Map<String, BigDecimal> getRateForSymbol(String symbol){
         return getRates().get(symbol);
     }
-    public static  Map<String, Map<String,Double>> getRates(){
-        Map<String, Map<String,Double>> rateData = new HashMap<>();
+    public static  Map<String, Map<String,BigDecimal>> getRates(){
+        Map<String, Map<String,BigDecimal>> rateData = new HashMap<>();
 
-        Map<String,Double> AEDrate = new HashMap<>();
-        AEDrate.put("AFN",24.774983);
-        AEDrate.put("ALL",31.185907);
-        AEDrate.put("AMD",124.553271);
-        AEDrate.put("ANG",0.490728);
-        AEDrate.put("AOA",112.099682);
-        AEDrate.put("AED",1.0);
+        Map<String,BigDecimal> AEDrate = new HashMap<>();
+        AEDrate.put("AFN", BigDecimal.valueOf(24.774983));
+        AEDrate.put("ALL", BigDecimal.valueOf(31.185907));
+        AEDrate.put("AMD", BigDecimal.valueOf(124.553271));
+        AEDrate.put("ANG", BigDecimal.valueOf(0.490728));
+        AEDrate.put("AOA", BigDecimal.valueOf(112.099682));
+        AEDrate.put("AED", BigDecimal.valueOf(1.0));
         rateData.put("AED",AEDrate);
 
-        Map<String,Double> AFNrate = new HashMap<>();
-        AFNrate.put("AED",0.040363);
-        AFNrate.put("ALL",1.258764);
-        AFNrate.put("AMD",5.027368);
-        AFNrate.put("ANG",0.019807);
-        AFNrate.put("AOA",4.524705);
-        AFNrate.put("AFN",1.0);
+        Map<String,BigDecimal> AFNrate = new HashMap<>();
+        AFNrate.put("AED", BigDecimal.valueOf(0.040363));
+        AFNrate.put("ALL", BigDecimal.valueOf(1.258764));
+        AFNrate.put("AMD", BigDecimal.valueOf(5.027368));
+        AFNrate.put("ANG", BigDecimal.valueOf(0.019807));
+        AFNrate.put("AOA", BigDecimal.valueOf(4.524705));
+        AFNrate.put("AFN", BigDecimal.valueOf(1.0));
         rateData.put("AFN",AFNrate);
 
-        Map<String,Double> ALLrate = new HashMap<>();
-        ALLrate.put("AED",0.032065);
-        ALLrate.put("AFN",0.79443);
-        ALLrate.put("AMD",3.993893);
-        ALLrate.put("ANG",0.015736);
-        ALLrate.put("AOA",3.594562);
-        ALLrate.put("ALL",1.0);
+        Map<String,BigDecimal> ALLrate = new HashMap<>();
+        ALLrate.put("AED", BigDecimal.valueOf(0.032065));
+        ALLrate.put("AFN", BigDecimal.valueOf(0.79443));
+        ALLrate.put("AMD", BigDecimal.valueOf(3.993893));
+        ALLrate.put("ANG", BigDecimal.valueOf(0.015736));
+        ALLrate.put("AOA", BigDecimal.valueOf(3.594562));
+        ALLrate.put("ALL", BigDecimal.valueOf(1.0));
         rateData.put("ALL",ALLrate);
 
 
-        Map<String,Double> AMDrate = new HashMap<>();
-        AMDrate.put("AED",0.008029);
-        AMDrate.put("AFN",0.198912);
-        AMDrate.put("ALL",0.250382);
-        AMDrate.put("ANG",0.00394);
-        AMDrate.put("AOA",0.900014);
-        AMDrate.put("AMD",1.0);
+        Map<String,BigDecimal> AMDrate = new HashMap<>();
+        AMDrate.put("AED", BigDecimal.valueOf(0.008029));
+        AMDrate.put("AFN", BigDecimal.valueOf(0.198912));
+        AMDrate.put("ALL", BigDecimal.valueOf(0.250382));
+        AMDrate.put("ANG", BigDecimal.valueOf(0.00394));
+        AMDrate.put("AOA", BigDecimal.valueOf(0.900014));
+        AMDrate.put("AMD", BigDecimal.valueOf(1.0));
         rateData.put("AMD",AMDrate);
 
-        Map<String,Double> ANGrate = new HashMap<>();
-        ANGrate.put("AED", 2.037759);
-        ANGrate.put("AFN", 50.486199);
-        ANGrate.put("ALL", 63.414802);
-        ANGrate.put("AMD", 253.813137);
-        ANGrate.put("AOA", 228.43548);
-        ANGrate.put("ANG",1.0);
+        Map<String,BigDecimal> ANGrate = new HashMap<>();
+        ANGrate.put("AED", BigDecimal.valueOf(2.037759));
+        ANGrate.put("AFN", BigDecimal.valueOf(50.486199));
+        ANGrate.put("ALL", BigDecimal.valueOf(63.414802));
+        ANGrate.put("AMD", BigDecimal.valueOf(253.813137));
+        ANGrate.put("AOA", BigDecimal.valueOf(228.43548));
+        ANGrate.put("ANG", BigDecimal.valueOf(1.0));
         rateData.put("ANG",ANGrate);
 
-        Map<String,Double> AOArate = new HashMap<>();
-        AOArate.put("AED", 0.008921);
-        AOArate.put("AFN", 0.22101);
-        AOArate.put("ALL", 0.277616);
-        AOArate.put("AMD", 1.111094);
-        AOArate.put("ANG", 0.004378);
-        AOArate.put("AOA",1.0);
+        Map<String,BigDecimal> AOArate = new HashMap<>();
+        AOArate.put("AED", BigDecimal.valueOf(0.008921));
+        AOArate.put("AFN", BigDecimal.valueOf(0.22101));
+        AOArate.put("ALL", BigDecimal.valueOf(0.277616));
+        AOArate.put("AMD", BigDecimal.valueOf(1.111094));
+        AOArate.put("ANG", BigDecimal.valueOf(0.004378));
+        AOArate.put("AOA", BigDecimal.valueOf(1.0));
         rateData.put("AOA",AOArate);
 
         return rateData;
